rename and shift of directories
