--- SHIFT_CIPHER.py
@@ -1,27 +1,0 @@
-"""
-   SHIFT CIPHER
-"""
-
-def encrypt(text,s):
-    result = ""
- 
-    # traverse text
-    for i in range(len(text)):
-        char = text[i]
- 
-        # Encrypt uppercase characters
-        if (char.isupper()):
-            result += chr((ord(char) + s-65) % 26 + 65)
- 
-        # Encrypt lowercase characters
-        else:
-            result += chr((ord(char) + s - 97) % 26 + 97)
- 
-    return result
- 
-#check the above function
-text = "HOW ARE YOU"
-s = 4
-print("Text  : " + text)
-print("Shift : " + str(s))
-print("Cipher: " + encrypt(text,s))
